GUARD-280: refactored code, simplified productsV3 service;
